部屋の概要欄のコンポーネント化
